added function to print rep as JSON to std out python fixes test server: repeat calls to retrieve /dev/pm as test

+++ api/oc_knx.c
@@ -1263,7 +1263,7 @@ oc_core_knx_spake_post_handler(oc_request_t *request,
     size_t device_index = request->resource->device;
     oc_device_info_t *device = oc_core_get_device_info(device_index);
     oc_oscore_set_auth(oc_string(device->serialnumber), oc_string(g_pase.id),
-                       shared_key, shared_key_len);
+                       shared_key, (int)shared_key_len);
 
     oc_send_cbor_response(request, OC_STATUS_CHANGED);
     // handshake completed successfully - clear state
